requirements.txt
Django==2.0.5
agavepy
pyjwt==1.6.4
